requirements.txt
Flask==3.0.3         
SQLAlchemy==2.0.32    
CoAPthon3==1.0.2
PyMySQL==1.1.1